requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2
django-filter==2.4.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
